# docs/swagger_v2.yml
swagger: '2.0'
info:
  title: fn
  description: The open source serverless platform.
  version: "2.0.0"
# the domain of the service
host: "127.0.0.1:8080"
...
schemes:
  - https
  - http
# will be prefixed to all paths
basePath: /v2
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /fns/{fnID}:
    get:
      operationId: "GetFn"
      summary: "Get definition for a function"
      description: "Get definition for a function."
      tags:
        - Fns
      parameters:
        - $ref: '#/parameters/FnID'
      responses:
        200:
          description: Function definition
          schema:
            $ref: '#/definitions/Fn'
        404:
          description: Function does not exist.
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
    put:
      operationId: "UpdateFn"
      summary: "Updates a fn"
      description: "Updates a Function via merging the provided values."
      tags:
        - Fns
      parameters:
        - $ref: '#/parameters/FnID'
        - name: body
          in: body
          description: Fn data to merge with current value
          required: true
          schema:
            $ref: '#/definitions/Fn'
      responses:
        200:
          description: Fn metadata
          schema:
            $ref: '#/definitions/Fn'
        404:
          description: Fn does not exist.
          schema:
            $ref: '#/definitions/Error'
        400:
          description: Parameters are missing or invalid.
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
    delete:
      operationId: "DeleteFn"
      summary: "Delete a fn"
      description: "Delete a function."
      tags:
        - Fns
      parameters:
        - $ref: '#/parameters/FnID'
      responses:
        204:
          description: Fn successfully deleted
        404:
          description: Fn does not exist.
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  Fn:
    type: object
    properties:
      id:
        type: string
        description: Unique identifier
        readOnly: true
      name:
        type: string
        description: unique name for this function.
      app_id:
        type: string
        description: App ID.
      image:
        type: string
        description: "full container image name, e.g. hub.docker.com/fnproject/yo or fnproject/yo (default registry: hub.docker.com)"
      mem:
        type: integer
        format: uint64
        description: Max usable memory given to function (MiB).
      timeout:
        type: integer
        default: 30
        format: int32
        description: Timeout for executions of a function. Value in Seconds
      idle_timeout:
        type: integer
        default: 30
        format: int32
        description: Hot functions idle timeout before container termination. Value in Seconds
      config:
        type: object
        description: Func configuration key values
        additionalProperties:
          type: string
      format:
        enum:
          - default
          - http
          - json
          - cloudevent
        description: Payload format sent into function.
        type: string
      annotations:
        type: object
        description: Func annotations - this is a map of annotations attached to this func, keys must not exceed 128 bytes and must consist of non-whitespace printable ascii characters, and the seralized representation of individual values must not exeed 512 bytes
        additionalProperties:
          type: object
      created_at:
        type: string
        format: date-time
        description: Time when function was created. Always in UTC RFC3339.
        readOnly: true
      updated_at:
        type: string
        format: date-time
        description: Most recent time that function was updated. Always in UTC RFC3339.
        readOnly: true
...
  ErrorBody:
    type: object
    properties:
      message:
        type: string
        readOnly: true
      fields:
        type: string
        readOnly: true
  Error:
    type: object
    properties:
      error:
        $ref: '#/definitions/ErrorBody'
parameters:
...
  FnID:
    name: fnID
    in: path
    description: Function ID.
    required: true
    type: string
...